connected the pages and assets together
